+++ core/src/main/java/com/scalar/db/storage/jdbc/RdbEngineOracle.java
@@ -1,8 +1,13 @@
 package com.scalar.db.storage.jdbc;
 
 import com.scalar.db.io.DataType;
+import org.apache.commons.dbcp2.BasicDataSource;
 
 class RdbEngineOracle extends RdbEngineStrategy {
+
+    RdbEngineOracle(BasicDataSource dataSource, RdbEngine rdbEngine, String metadataSchema) {
+        super(dataSource, rdbEngine, metadataSchema);
+    }
 
     @Override
     String getDataTypeForEngine(DataType scalarDbDataType) {
